+++ lib/twenv/command_mixins/tweet_filter.rb
@@ -10,4 +10,14 @@ module TWEnv::Command::TweetFilter
       filtered
     end
   end
+
+  def is_reply_to_filter!(tweets)
+    tweets.select! do |tweet|
+      next unless tweet.reply?
+      user = client.user(tweet.in_reply_to_user_id)
+      user.screen_name == opts['is-reply-to']
+    rescue Twitter::Error::NotFound, Twitter::Error::Forbidden
+      nil
+    end if opts['is-reply-to']
+  end
 end