+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD
 
+* Add `--is-reply-to=` option to `archive-{timeline,likes}`.
+`
 * Add `Back to top` links to the README.
 
 * Add `LICENSE.txt`.